🚚 Move Stars into their own component
